rename files and run dos2unix
